0.0.15
Triangulation implementation done

+++ openCVImplementation/JMPRcheckRot.m
@@ -11,7 +11,7 @@
 @implementation JMPRcheckRot
 
 
-bool checkrot(Mat_<double> &R){
+bool checkrot(Mat_<float> &R){
     if(int(fabsf(determinant(R)-1.0 > 1e-07))) {
         return false;
         cout<<"This is not a correct rotation \n";
